+++ GAC-EJBSession/ejbModule/local/ConversationServiceLocal.java
@@ -9,7 +9,7 @@ public interface ConversationServiceLocal {
 
 	  public void addItem(Conversation i);
 
-	    public Conversation findItem(long id);
+	    public Conversation findItem(Integer id);
 
 	    public void deleteItem(Conversation i);
